day151/gitops-log-platform/requirements.txt
# GitOps Controller Dependencies (May 2025 compatible)
kubernetes==29.0.0
gitpython==3.1.43
pyyaml==6.0.1
fastapi==0.111.0
uvicorn==0.30.0
pydantic==2.7.1
jinja2==3.1.4
aiohttp==3.9.5
prometheus-client==0.20.0
python-multipart==0.0.9
structlog==24.1.0
watchdog==4.0.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-cov==5.0.0
httpx==0.27.0

# Development
black==24.4.2
flake8==7.0.0
